requirements.txt
discord.py[audio]
